requirements.txt
anthropic>=0.40.0
tavily-python>=0.5.0
trafilatura>=1.12.0
requests>=2.31.0
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0